requirements.txt
autopep8==1.6.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.5
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.0.2
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-vision==3.1.1
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
platformdirs==2.5.2
proto-plus==1.22.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.9
six==1.16.0
toml==0.10.2
tqdm==4.62.3
twilio==7.12.1
types-pytz==2022.1.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.16.3
vonage==2.5.5
Werkzeug==2.0.2
wrapt==1.13.3
